Moving filter policy related logic out of utils and into subscription filter policies.
