+++ src/main/kotlin/com/jameskbride/localsns/models/MessageBodyFilterPolicy.kt
@@ -1,10 +1,9 @@
 package com.jameskbride.localsns.models
 
 import com.google.gson.Gson
-import com.google.gson.JsonElement
 import com.google.gson.JsonObject
-import com.jameskbride.localsns.getElementType
-import com.jameskbride.localsns.validateNumericMatcher
+import com.jameskbride.localsns.subscriptions.getElementType
+import com.jameskbride.localsns.subscriptions.validateNumericMatcher
 
 class MessageBodyFilterPolicy(filterPolicy: String) {
     private val gson = Gson()
